Moved filter typehints to their own dedicated module
